move ControlLayer into SettingHUDOverlay
