+++ osu.Game.Rulesets.Karaoke/UI/Overlays/Settings/RightSideOverlay.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics.Containers;
 using osuTK;
 using osuTK.Graphics;
-using static osu.Game.Rulesets.Karaoke.UI.HUD.ControlLayer;
+using static osu.Game.Rulesets.Karaoke.UI.Overlays.SettingHUDOverlay;
 
 namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings
 {
